refactor: disable pg log
