+++ .env.example
@@ -1,4 +1,5 @@
 NODE_PORT=4000
+NODE_ENV=production
 
 POSTGRES_PORT=5432
 POSTGRES_HOST=postgres
