--- src/auth/interfaces/auth-answer.interface.ts
@@ -1,4 +1,0 @@
-export interface IAuthAnswer {
-  accessToken: string;
-  refreshToken: string;
-}
