+++ src/auth/interfaces/auth.interfaces.ts
@@ -1,0 +1,8 @@
+interface IAuthAnswer {
+  accessToken: string;
+  refreshToken: string;
+}
+
+type RefreshPayload = { id: string; type: string };
+
+export { IAuthAnswer, RefreshPayload };
