+++ src/utils/public.ts
@@ -1,4 +1,4 @@
 import { SetMetadata } from '@nestjs/common';
+import { IS_PUBLIC_KEY } from './constants';
 
-export const IS_PUBLIC_KEY = 'isPublic';
 export const Public = () => SetMetadata(IS_PUBLIC_KEY, true);
